django-blog app/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.2
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
